Fix broken concatenated strings
A string that is just a fullstop!

If these are not an example of why concatenated strings are evil I don't know
what is.

Will be using this example in my training to programmers on why concatenated
strings are bad bad bad.  Fix the program don't make an already hard task of
localisation harder! </rant>